refactor(shopping): 패키지 구조 분리
