+++ app/src/main/java/woowacourse/shopping/shopping/adapter/viewholder/ShoppingRecyclerItemViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.shopping
+package woowacourse.shopping.shopping.adapter.viewholder
 
 import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
